rename CXX to IR closes #5
