--- capdpa/CXX/__init__.py
@@ -1,9 +1,0 @@
-
-from .cxx_namespace import Namespace
-from .cxx_class import Class
-from .cxx_function import Function
-from .cxx_type import Type
-from .cxx_variable import Variable
-from .cxx_constant import Constant
-from .cxx_enum import Enum
-from .cxx_identifier import Identifier
