+++ capdpa/IR/__init__.py
@@ -1,0 +1,9 @@
+
+from .ir_namespace import Namespace
+from .ir_class import Class
+from .ir_function import Function
+from .ir_type import Type
+from .ir_variable import Variable
+from .ir_constant import Constant
+from .ir_enum import Enum
+from .ir_identifier import Identifier
